python-grpc_lab/requirements.txt
text
grpcio==1.47.0
grpcio-tools==1.47.0
protobuf==3.20.1